PG ~ Add documentations and connect analyzer script to ProcWatcher
