+++ ProcGuard/src/include/common.h
@@ -7,7 +7,8 @@ enum class ProcStatus : uint8_t
     OK = 0,
     WAITING = 1,
     CHECKING = 2,
-    WARNING = 3,
+    SUSPICIOUS = 3,
+    ANALYZE_ERROR = 4,
     NOT_EXISTS = 0xFF,
 };
