+++ ProcGuard/strace_parser.py
